requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.4.0
ffmpeg-python
imageio-ffmpeg>=0.6.0

pydub>=0.25.1
aiofiles>=23.2.1
jinja2>=3.1.2
python-dotenv>=1.0.0
